requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
et-xmlfile==1.0.1
feedfinder2==0.0.4
feedparser==6.0.1
filelock==3.0.12
GoogleNews==1.4.1
h5py==2.10.0
idna==2.10
jdcal==1.4.1
jieba3k==0.35.1
joblib==0.16.0
jsonlines==1.2.0
Keras==2.4.3
lxml==4.5.2
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.2
olefile==0.46
openpyxl==3.0.5
packaging==20.4
pandas==1.1.2
Pillow @ file:///C:/ci/pillow_1594298230227/work
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-file==1.5.1
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.2
sentencepiece==0.1.91
seqeval==0.0.12
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==2.2.3
tokenizers==0.8.1rc2
torch==1.6.0
torchvision==0.7.0
tqdm==4.49.0
transformers==3.1.0
urllib3==1.25.10
wincertstore==0.2
xlrd==1.2.0
XlsxWriter==1.3.3
